Hidden bug fixed
Changed the way how brick should be destroyed.

When mario jumps very slow, he gives the world.step method the chance to actually calculate the head filter and body filter in two different iterations. The way how brick was destroyed was depending heavily on the order of the collision detection. If the head was detected first and then the body, the body wouldnt be stopped by the collision since the filter bit was changed already.